fix: Agent Backups

+++ server/server_agent_handler.js
@@ -87,15 +87,17 @@ class AgentHandler {
             for (let j = 0; j < count; j++) {
                 const AgentBackup = Agent.backups[j];
 
-                if (fs.existsSync(AgentBackup.fileName)) {
-                    fs.unlinkSync(AgentBackup.fileName);
-                    Logger.debug(
-                        `Removing Agent Backup: ${AgentBackup.fileName}`
-                    );
+                if(AgentBackup != null && AgentBackup.fileName != null){
+                    if (fs.existsSync(AgentBackup.fileName)) {
+                        fs.unlinkSync(AgentBackup.fileName);
+                        Logger.debug(
+                            `Removing Agent Backup: ${AgentBackup.fileName}`
+                        );
+                    }
+    
+                    await AgentBackupModel.deleteOne({ _id: AgentBackup._id });
+                    Agent.backups.pull({ _id: AgentBackup._id });
                 }
-
-                await AgentBackupModel.deleteOne({ _id: AgentBackup._id });
-                Agent.backups.pull({ _id: AgentBackup._id });
             }
 
             await Agent.save();
